Move ConsoleKit to PPNetLib
